Consolidate dynamo calls
