+++ packages/core/src/utils/types.ts
@@ -1,0 +1,3 @@
+export type User = {
+  user_id: string;
+};
